Generate an Array of All Object Keys with Object.keys()
We can also generate an array which contains all the keys stored in an object using the Object.keys() method and passing in an object as the argument. This will return an array with strings representing each property in the object. Again, there will be no specific order to the entries in the array.


Finish writing the getArrayOfUsers function so that it returns an array containing all the properties in the object it receives as an argument.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/keys